fix: refresh trace compare-variant against main
